+++ src/db/model/mod.rs
@@ -1,3 +1,4 @@
 pub mod playlist;
+pub mod transfer;
 pub mod user;
 pub mod watcher;