+++ src/db/repo/mod.rs
@@ -1,2 +1,3 @@
+pub mod transfer;
 pub mod user;
 pub mod watcher;