+++ src/sync/mod.rs
@@ -1,10 +1,10 @@
 use self::error::SyncResult;
 use crate::{
-    api::client,
+    api::client::{self, Client},
     context::AppContext,
     db::{
         model::watcher::Watcher,
-        repo::{user::UserRepo, watcher::WatcherRepo},
+        repo::{transfer::TransferRepo, user::UserRepo, watcher::WatcherRepo},
     },
     sync::error::SyncError,
 };
@@ -12,9 +12,6 @@ use chrono::{DateTime, Timelike, Utc};
 
 pub mod error;
 pub mod transfer;
-
-/// Interval to fetch watchers to see if any need to be run again
-const CHECK_INTERVAL_MINS: i64 = 1;
 
 pub async fn init(ctx: AppContext) -> SyncResult<()> {
     loop {
@@ -24,7 +21,9 @@ pub async fn init(ctx: AppContext) -> SyncResult<()> {
             .unwrap()
             .with_nanosecond(0)
             .unwrap()
-            .checked_add_signed(chrono::Duration::minutes(CHECK_INTERVAL_MINS))
+            .checked_add_signed(chrono::Duration::minutes(
+                ctx.config.sync.check_interval_mins.into(),
+            ))
             .unwrap();
 
         let time_until_next_update = tokio::time::Duration::from_millis(
@@ -33,7 +32,7 @@ pub async fn init(ctx: AppContext) -> SyncResult<()> {
 
         tracing::info!(
             "{:.0}s until next check",
-            time_until_next_update.as_secs_f64()
+            time_until_next_update.as_secs_f64(),
         );
 
         tokio::time::sleep_until(
@@ -68,11 +67,26 @@ async fn execute(ctx: AppContext) -> SyncResult<()> {
     let now = Utc::now().with_second(0).unwrap().with_nanosecond(0).unwrap();
 
     for watcher in to_sync {
-        if let Err(err) = sync_watcher(ctx.clone(), &user_repo, &watcher_repo, &watcher, now).await
-        {
-            // Don't kill worker thread if an individual sync task errored
-            tracing::error!("Error when syncing watcher: {}", err);
-            sentry::capture_error(&err);
+        let user = user_repo
+            .find_user_by_uri(&watcher.user_uri)?
+            .ok_or_else(|| SyncError::UserNotFoundError(watcher.user_uri.clone()))?;
+
+        let client = client::Client::new_with_token(ctx.clone(), user.token)?;
+        client.ensure_token_refreshed(&watcher.user_uri).await?;
+
+        match sync_watcher(ctx.clone(), client, &watcher_repo, &watcher, now).await {
+            Ok(_) => {
+                // Set the next interval
+                watcher_repo.update_watcher_next_sync_at(
+                    watcher.id,
+                    now.checked_add_signed(watcher.sync_interval.clone().into()).unwrap(),
+                )?;
+            }
+            Err(err) => {
+                // Don't kill worker thread if an individual sync task errored
+                tracing::error!("Error when syncing watcher: {}", err);
+                sentry::capture_error(&err);
+            }
         }
     }
 
@@ -81,29 +95,39 @@ async fn execute(ctx: AppContext) -> SyncResult<()> {
     Ok(())
 }
 
-async fn sync_watcher(
+/// Sync a watcher and save the results to the transfer table.
+pub async fn sync_watcher(
     ctx: AppContext,
-    user_repo: &UserRepo,
+    client: Client,
     watcher_repo: &WatcherRepo,
     watcher: &Watcher,
     now: DateTime<Utc>,
-) -> SyncResult<()> {
-    let user = user_repo
-        .find_user_by_uri(&watcher.user_uri)?
-        .ok_or_else(|| SyncError::UserNotFoundError(watcher.user_uri.clone()))?;
+) -> SyncResult<u32> {
+    let res = sync_watcher_inner(ctx.clone(), client, &watcher_repo, watcher, &now).await;
 
-    let client = client::Client::new_with_token(ctx.clone(), user.token)?;
-    client.ensure_token_refreshed(&watcher.user_uri).await?;
-
-    transfer::PlaylistTransfer::new(ctx.clone(), client)
-        .try_transfer(&watcher)
-        .await?;
-
-    watcher_repo.update_watcher_last_sync_at(watcher.id, now)?;
-    watcher_repo.update_watcher_next_sync_at(
+    // Save transfer result
+    TransferRepo::new(ctx.clone()).log_transfer(
         watcher.id,
-        now.checked_add_signed(watcher.sync_interval.clone().into()).unwrap(),
+        res.as_ref().unwrap_or(&0),
+        &res.as_ref().err(),
+        now,
     )?;
 
-    Ok(())
+    res
+}
+
+/// Sync a watcher and update the `last_sync_at` date
+async fn sync_watcher_inner(
+    ctx: AppContext,
+    client: Client,
+    watcher_repo: &WatcherRepo,
+    watcher: &Watcher,
+    now: &DateTime<Utc>,
+) -> SyncResult<u32> {
+    let num_tracks_transferred =
+        transfer::PlaylistTransfer::new(ctx, client).try_transfer(&watcher).await?;
+
+    watcher_repo.update_watcher_last_sync_at(watcher.id, *now)?;
+
+    Ok(num_tracks_transferred)
 }